LayoutTests/platform/gtk/imported/w3c/web-platform-tests/html/semantics/embedded-content/the-iframe-element/iframe_sandbox_navigation_download_allow_downloads.sub.tentative-expected.txt

CONSOLE MESSAGE: line 1984: Error while parsing the 'sandbox' attribute: 'allow-downloads' is an invalid sandbox flag.
CONSOLE MESSAGE: Blocked script execution in 'about:srcdoc' because the document's frame is sandboxed and the 'allow-scripts' permission is not set.


FAIL Navigation resulted download in sandbox is allowed by allow-downloads. undefined is not an object (evaluating 'anchor.href = "support/download_stash.py?token=" + token + "&finish-delay=" + StreamDownloadFinishDelay()')
